refactor(cli): refactor bmp functions out of cli
this allows for additional file formats to be added in the future

resolves #12
